feat:新增角色表 新增权限表 新增用户和角色的关系表 新增角色和权限的关系表 新增数据库初始化权限角色以及对应关系的数据填充

+++ Migrations/20260224081501_AddUserRole.cs
@@ -15,7 +15,7 @@ namespace day1.Migrations
                 name: "Users",
                 columns: table => new
                 {
-                    Id = table.Column<int>(type: "int", nullable: false)
+                    Id = table.Column<long>(type: "bigint", nullable: false)
                         .Annotation("SqlServer:Identity", "1, 1"),
                     UserName = table.Column<string>(type: "nvarchar(max)", nullable: false),
                     PassWord = table.Column<string>(type: "nvarchar(max)", nullable: false),
